requirements.txt
pyyaml==5.1.2
openapi_core==0.11.0
jinja2==2.10.1
aiohttp==3.5.4
python-rapidjson==0.8.0